Move pyam-related computations into reporting.pyam
