+++ message_ix/reporting/pyam.py
@@ -1,7 +1,9 @@
 from logging import getLogger
 
+from ixmp.reporting.computations import write_report as ixmp_write_report
+from ixmp.reporting.utils import concat as ixmp_concat
 import pandas as pd
-from pyam import IAMC_IDX, IamDataFrame
+from pyam import IAMC_IDX, IamDataFrame, concat as pyam_concat
 
 
 log = getLogger(__name__)
@@ -57,6 +59,36 @@ def as_pyam(scenario, year_time_dim, quantities, drop=[], collapse=None):
                     .format(extra, [q.name for q in quantities]))
 
     return IamDataFrame(df)
+
+
+# Computations that operate on pyam.IamDataFrame inputs
+
+def concat(*args):
+    """Concatenate *args*, which must be :class:`pyam.IamDataFrame`."""
+    if isinstance(args[0], IamDataFrame):
+        return pyam_concat(args)
+    else:
+        return ixmp_concat(args)
+
+
+def write_report(quantity, path):
+    """Write the report identified by *key* to the file at *path*.
+
+    If *quantity* is a :class:`pyam.IamDataFrame` and *path* ends with '.csv'
+    or '.xlsx', use :mod:`pyam` methods to write the file to CSV or Excel
+    format, respectively. Otherwise, equivalent to
+    :meth:`ixmp.reporting.computations.write_report`.
+    """
+    if not isinstance(quantity, IamDataFrame):
+        return ixmp_write_report(quantity, path)
+
+    if path.suffix == '.csv':
+        quantity.to_csv(path)
+    elif path.suffix == '.xlsx':
+        quantity.to_excel(path, merge_cells=False)
+    else:
+        raise ValueError('pyam.IamDataFrame can be written to .csv or .xlsx, '
+                         'not {}'.format(path.suffix))
 
 
 def collapse_message_cols(df, var, kind=None):
